Ability to pass root and placeholder elements
This is a way to allow users to pass a rootElement of any html element like tr or span. By default it uses div. 
Also allows you to provide a placeholderElement of any html element like tr or span. By default it uses div.

This allows people that are trying to use this with tables to provide the type of html element they want the root and/or placeholder elements to be so it complies with valid html.